The twelve content failures were one trim-rule defect, not twelve tests #234
Triage of the twelve, classified before any was touched. All twelve are
property assertions. None is updated; not one assertion changes in this
commit. Each failed because the re-taken archive was cut for a reader it
does not serve, and the fix is in the capture's trim rule and in the shim
that serves it.

  1-2  the_rewritten_week_does_reach_the_features... [Lamb, Robinson]
       PROPERTY: the leakage test's premise, that a rewrite reaches later
       weeks. Failed because the screen Panel had 13 rows.
  3    the_trend_features_have_something_to_compute_on
       PROPERTY: the archive's floor, in numbers. 13 non-null of 13.
  4    the_panel_carries_the_designation_and_the_out_weeks_do_not_reach_it
       PROPERTY: Out never reaches a row, Questionable does. 0 of 9 reached
       because the 13-row Panel held none of the injury weeks.
  5    the_gate_spec_keeps_the_players_the_consensus_page_leaves_out
       PROPERTY: consensus=False is the wider Panel. 410 and 344 are the
       cohort over the page's own coverage and both hold again unchanged.
  6    the_expected_spec_moves_the_priors_and_not_the_realised_columns
       PROPERTY. Failed in the shim: ff_opportunity ships season as a
       string, the capture records that, and by_season compared it to ints.
  7-8  a_total_on_the_panel_is_the_sum... / the_scoring_prior_splits... [expected]
       PROPERTY: arithmetic identities. Same shim failure as 6.
  9    the_touchdown_rate_prior_divides_one_measurement_by_itself
       PROPERTY. Same shim failure as 6.
  10   every_column_the_panel_hands_back_is_on_exactly_one_side [expected]
       PROPERTY: the default-deny at _served. Same shim failure as 6.
  11   a_dome_a_missing_reading_and_a_measured_calm_are_three_different_things
       PROPERTY: #170's distinction. The edit reached no row of a 13-row Panel.
  12   the_routed_sources_really_go_through_the_loader_when_a_panel_is_built
       PROPERTY: the behavioural half of the routing. Same shim failure as 6.

What the archive showed. The first run of the script cut ff_rankings to
the draft board's preseason window, a reader this archive does not serve
(draft_board.json serves it), and took "the last four dates starting with
2023" from that window, which is December 2023: eight scrape dates, none
of them a week the Panel's weekly_consensus assigns to weeks 1-14, so the
screen's inner join kept 13 rows of 344. It matched the cohort by display
name, so Michael Pittman Jr. and Patrick Mahomes II fell out of the
sources that spell them so. And it left ff_opportunity at 409 rows under a
contract whose floor is 1,000; nflverse.load refuses it, and the script had
never asked.

What changed. scripts/capture_panel_archive.py: the cohort is matched on
player_key; ff_rankings keeps the weekly-op page at every scrape the
Panel's own assign_weeks maps onto weeks 1-14 of either season;
ff_opportunity is padded to its contract's floor from the same cells in
(season, week, player_id) order, the rule the README already stated; and
every capture is validated against its contract before it is written.
tests/panelarchive.py: by_season filters on the integer the argument
names, as nflreadpy's per-season load does regardless of the column.

Two files re-taken offline, recorded in the README. ff_rankings and
ff_opportunity were re-taken through the corrected capture() from the
repo's pinned cache entries (data/raw/nflverse/ff_rankings/all-fadb12e5,
ff_opportunity/{2023,2024}-5ef5ef03: real frames nflverse.load fetched,
validated and pinned on 2026-09-06/07). ff_rankings is append-only and both
seasons are closed, so a live --write returns the same rows; the 365 are
byte-for-byte the 2026-09-05 hand capture's rows on those dates, the other
91 being week-15+ scrapes that join to no Panel row. snap_counts could not
be re-taken (no cache entry): Pittman Jr.'s 25 rows stay absent until
--write is run again. A new test states the cohort property on player_key
across the per-player captures, with that one case a strict xfail that
flips the day the archive is re-taken. Mutant killed: Tyreek Hill renamed
throughout ff_rankings.json fails the ff_rankings case.

Gates: ruff 0, pyrefly 0, pytest 754 passed, 1 xfailed, 3 failed. The three
are test_the_three_unrouted_sources_are_blocked_by_the_archive_and_not_by_
this_module, the designed signal that the archive no longer blocks; they
are #35's and the next commit replaces them.

+++ tests/panelarchive.py
@@ -104,7 +104,12 @@ def install(monkeypatch, tmp_path: Path, *, edits: dict[str, Edit] | None = None
         def load(seasons=None, **_kw):
             df = served(name)
             want = _seasons(seasons)
-            return df if want is None else df.filter(pl.col("season").is_in(want))
+            # nflreadpy picks the season *files* to read, so its `seasons` argument never
+            # meets the column's dtype. This filter does, and `ff_opportunity` ships `season`
+            # as a string -- the capture records that, as it should, and `expected_weekly`
+            # casts it -- so the comparison is made on the integer the argument names.
+            return df if want is None else df.filter(
+                pl.col("season").cast(pl.Int64).is_in(want))
         return load
 
     def refuse(source: str):
